requirements.txt
requests==2.32.3
python-telegram-bot==20.0
fastapi== 0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35


